Refactor trip creation components and update imports
- Updated import paths for UnifiedTripCreator in create-ai-trip.tsx and create-manual-trip.tsx to reflect new directory structure.
- Removed unused components including CreateTripNavigationBar, DayPlanner, FoodSpotType, and others to streamline the codebase.
- Added utility functions for date formatting and calculation in Datetime.ts.
- Enhanced trip detail view by integrating useTripDetails hook for better state management.
