+++ app/(tabs)/my-trips/create-ai-trip.tsx
@@ -1,5 +1,5 @@
 import { createAiTripSteps, TRIP_TYPES } from '@/constants/createTripSteps'
-import UnifiedTripCreator from '@/features/trip/presentation/components/UnifiedTripCreator'
+import UnifiedTripCreator from '@/features/trip/presentation/components/CreateTrip/UnifiedTripCreator'
 import { useThemeStyle } from '@/hooks/useThemeStyle'
 import { useAiTripStore } from '@/store/useAiTripStore'
 import { useFocusEffect } from '@react-navigation/native'
