+++ components/LoadingScreen.tsx
@@ -1,0 +1,32 @@
+import { colorPalettes } from '@/constants/Itheme'
+import { useThemeStyle } from '@/hooks/useThemeStyle'
+import React, { useMemo } from 'react'
+import { SafeAreaView, StatusBar, StyleSheet, Text, View } from 'react-native'
+
+const LoadingScreen: React.FC = () => {
+  const theme = useThemeStyle()
+  const styles = useMemo(() => createStyles(theme), [theme])
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <StatusBar barStyle="dark-content" />
+      <View style={styles.content}>
+        <Text>Loading trip details...</Text>
+      </View>
+    </SafeAreaView>
+  )
+}
+
+const createStyles = (theme: typeof colorPalettes.light) =>
+  StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: theme.white,
+    },
+    content: {
+      flex: 1,
+      marginTop: 20,
+    },
+  })
+
+export default LoadingScreen
