+++ constants/createTripSteps.ts
@@ -1,14 +1,14 @@
-import FoodSpotType from '@/features/trip/presentation/components/FoodSpotType'
-import LocPreference from '@/features/trip/presentation/components/LocPreference'
-import ManualTripCreate from '@/features/trip/presentation/components/ManualTripCreate'
-import MedicalReqs from '@/features/trip/presentation/components/MedicalReq'
-import OtherReqs from '@/features/trip/presentation/components/OtherReq'
-import SpotNumber from '@/features/trip/presentation/components/SpotNumber'
-import SpotType from '@/features/trip/presentation/components/SpotType'
-import TripLength from '@/features/trip/presentation/components/TripLength'
-import ChooseLocation from '@/features/trip/presentation/components/TripLocation'
-import TripRename from '@/features/trip/presentation/components/TripRename'
-import WaitScreen from '@/features/trip/presentation/components/WaitScreen'
+import FoodSpotType from '@/features/trip/presentation/components/CreateTrip/AI/FoodSpotType'
+import LocPreference from '@/features/trip/presentation/components/CreateTrip/AI/LocPreference'
+import MedicalReqs from '@/features/trip/presentation/components/CreateTrip/AI/MedicalReq'
+import OtherReqs from '@/features/trip/presentation/components/CreateTrip/AI/OtherReq'
+import ManualTripCreate from '@/features/trip/presentation/components/CreateTrip/ManualTripCreate'
+import SpotNumber from '@/features/trip/presentation/components/CreateTrip/SpotNumber'
+import SpotType from '@/features/trip/presentation/components/CreateTrip/SpotType'
+import TripLength from '@/features/trip/presentation/components/CreateTrip/TripLength'
+import ChooseLocation from '@/features/trip/presentation/components/CreateTrip/TripLocation'
+import TripRename from '@/features/trip/presentation/components/CreateTrip/TripRename'
+import WaitScreen from '@/features/trip/presentation/components/CreateTrip/WaitScreen'
 import { ComponentType } from 'react'
 
 export const TRIP_TYPES = {
