+++ features/trip/presentation/components/CreateTrip/CreateTripNavigationBar.tsx
