+++ features/trip/presentation/components/CreateTrip/TripItemCard.tsx
