+++ features/trip/presentation/components/CreateTrip/TripLocation.tsx
